backend/users/serializers.py
from rest_framework import serializers
from .models.custom_user import CustomUser


class CustomUserSerializer(serializers.ModelSerializer):
    class Meta:
        model = CustomUser
        fields = '__all__'
