backend/requirements-prod.txt
appdirs==1.4.4
distlib==0.3.1
filelock==3.0.12
six==1.15.0
virtualenv==20.4.2
Django==3.1.7
Pillow==8.1.2
psycopg2==2.8.6
drf-spectacular==0.13.2
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
jsonschema==3.2.0
django-filter==2.4.0